Move Hotspot Private Methods Into Static Util Public Methods
In an effort to make code more testable and scalable, begin by
moving hotspot methods out into a static util class with public
methods.
